Fix memory leak that has forever been occurring in prod
The 'serverless' target would misbehave with a normal containerized
deployment in EC2. On the first time a unique page was loaded it would
take a new 28MB chunk from it into memory and hold it there seemingly
indefinitely. This would cause the memory usage to spike up over time a
lot, basically without a bound. With this change, the memory usage of
the production server locally would seem to be at around 200-300MB, even
with moderate use.