Fix table pagination throwing an error on every /[id] page
The reason for this was that `<PaginatedTable />` uses
`getQueryWithPagination()`, which only takes those key/value pairs from
`query` that are `'page'` or `'pageSize'` or passed in `extraFilters`.
This made it so that the call on `Router.push()` would not get the
`'id'` param that the page requires and it'd throw this error:
https://github.com/vercel/next.js/blob/master/errors/href-interpolation-failed.md

To fix this, on every `/[id]` page we pass the whole current query to
the component so that `'id'` for sure gets passed to `Router.push()`.

This implementation works, but might not be optimal since
`getQueryWithPagination()` doesn't actually need the the values of the
passed `extraFilters` at all, but only checks that the keys exist. Also
we are now reduadantly passing `'page'` and `'pageSize'` also as
`extraFilters={variables}`, even though we only need the `'id'` key there.