Don't hide the error details if fetching the sitemap fails
This error details won't be exposed to the client anyways, since this is
rendered completely on the server, but it's just easier to see the error
immediately from the server logs. The client will also actually see the
proper 500 page now, instead of getting an empty page with just the
status code.