Sort blogs and product updates in inverse chronological order
Tried to make the dates be part of the filenames and then just sort the
filenames, but it didn't work on `blogs/[slug].tsx` page's
`getStaticProps`, which needs to just open the files based on the URL
slug and doesn't know anything about the dates.

Also `BlogsPage` and `UpdatesPage` components are atm very similar and could
surely be refactor to use some shared functionality, but since the code
is only in in two places, it's okayish for now. Semi-relevant:
https://sandimetz.com/blog/2016/1/20/the-wrong-abstraction