Use the later of BUILD_DATE or `obj.modified` as sitemap's <lastmod>
See discussion in:
https://skoleinc.slack.com/archives/CMM0BSW9J/p1612772433001300

We also don't need to take the date from the blog files or updates
anymore, since the build date is always a later date and we want to just
use that. The HTML - and especially the meta tags of a page - can
easily change even if the markdown file would not be changed.