Fix some console warnings
The `SettingsButton` component already includes the tooltip so it's not
needed there.

Still getting e.g.:
```
The value provided to the Tabs component is invalid.
None of the Tabs' children match with `undefined`.

validateDOMNesting(...): <div> cannot appear as a child of <tr>.

validateDOMNesting(...): <td> cannot appear as a child of <div>.
```

But at least the DOM nesting ones in `CourseTableRow` component seemed
really hard to fix.