Use server-side rendering on dynamic pages to improve SEO
This allows the SEO meta tags to be populated immediately with the
correct data in the pre-rendered HTML. Very useful since Facebook's
scraper does not execute any JS.

Use `getServerSideProps`, instead of `getStaticProps` with a short
`revalidate` time, so that if a user updates their profile they will see
the info on the first load of their profile page, and not on the second.

This also now returns the `notFound` property which makes the not found
pages actually return 404 status code. This is strongly encouraged by
for example Google as well.